+++ src/RestSharp/Authenticators/OAuth2/OAuth2AuthorizationRequestHeaderAuthenticator.cs
@@ -28,14 +28,8 @@ public class OAuth2AuthorizationRequestHeaderAuthenticator : AuthenticatorBase {
     /// Initializes a new instance of the <see cref="OAuth2AuthorizationRequestHeaderAuthenticator" /> class.
     /// </summary>
     /// <param name="accessToken">The access token.</param>
-    public OAuth2AuthorizationRequestHeaderAuthenticator(string accessToken) : this(accessToken, "OAuth") { }
-
-    /// <summary>
-    /// Initializes a new instance of the <see cref="OAuth2AuthorizationRequestHeaderAuthenticator" /> class.
-    /// </summary>
-    /// <param name="accessToken">The access token.</param>
     /// <param name="tokenType">The token type.</param>
-    public OAuth2AuthorizationRequestHeaderAuthenticator(string accessToken, string tokenType) : base(accessToken) => _tokenType = tokenType;
+    public OAuth2AuthorizationRequestHeaderAuthenticator(string accessToken, string tokenType = "OAuth") : base(accessToken) => _tokenType = tokenType;
 
     protected override ValueTask<Parameter> GetAuthenticationParameter(string accessToken)
         => new(new HeaderParameter(KnownHeaders.Authorization, $"{_tokenType} {accessToken}"));